requirement.txt
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
cryptography==3.4.7
decorator==5.0.9
Django==3.2.5
django-appconf==1.0.4
django-crispy-forms==1.12.0
django-extensions==3.1.3
django-paypal==1.1.1
django-responsive2==0.1.3
django-shell-plus==1.1.7
djangorestframework==3.12.4
idna==2.10
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
Naked==0.1.31
parso==0.8.2
paypal-checkout-serversdk==1.0.1
paypalhttp==1.0.0
paypalrestsdk==1.13.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.1
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycparser==2.20
pycryptodome==3.10.1
Pygments==2.9.0
pyOpenSSL==20.0.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
shellescape==3.8.1
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
